refactor and file movement
