+++ apps/src/signUpFlow/FreeCurriculumDialog.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import Button from '@cdo/apps/componentLibrary/button/Button';
 import {Heading1, BodyThreeText} from '@cdo/apps/componentLibrary/typography';
 import AccessibleDialog from '@cdo/apps/sharedComponents/AccessibleDialog';
-import locale from '@cdo/apps/signup/locale';
+import locale from '@cdo/apps/signUpFlow/locale';
 
 import SafeMarkdown from '../templates/SafeMarkdown';
