+++ apps/src/signUpFlow/loginTypeSelection.module.scss
@@ -1,10 +1,16 @@
 @import "color";
 
-#newSignupFlow {
+body {
+  background-color: $light_gray_50 !important;
+}
+
+.newSignupFlow {
   display: flex;
   flex-direction: column;
   align-items: center;
   text-align: center;
+  padding-block: 1rem;
+  margin-inline: auto;
 
   .containerWrapper {
     display: flex;
@@ -24,23 +30,23 @@
     flex-direction: column;
     align-items: flex-start;
     text-align: left;
+
+    .greenButton {
+      background-color: $google_brand_color;
+    }
+  
+    .yellowButton {
+      background-color: $microsoft_brand_color;
+      color: $black;
+    }
+  
+    .blueButton {
+      background-color: $facebook_brand_color;
+    }
   }
 
   .container > * {
     margin-bottom: 20px; /* Adds space between each item */
     width: 100%;
-  }
-
-  .greenButton {
-    background-color: $google_brand_color;
-  }
-
-  .yellowButton {
-    background-color: $microsoft_brand_color;
-    color: $black;
-  }
-
-  .blueButton {
-    background-color: $facebook_brand_color;
   }
 }
