+++ dashboard/app/views/devise/registrations/_new_sign_up_locale.html.haml
@@ -1,0 +1,2 @@
+- js_locale = request.locale.to_s.downcase.tr('-', '_')
+%script{src: webpack_asset_path("js/#{js_locale}/signup_locale.js"), 'data-ot-ignore' => ''}
