+++ dashboard/app/views/devise/registrations/account_type.html.haml
@@ -1,3 +1,7 @@
+- content_for :head do
+  = render 'devise/registrations/new_sign_up_locale'
+
 %script{src: webpack_asset_path('js/devise/registrations/account_type.js')}
 #account-type
+  = render 'devise/registrations/new_sign_up_locale'
   -# Populated by React
