+++ dashboard/app/views/devise/registrations/new_sign_up.html.haml
@@ -1,4 +1,7 @@
+- content_for :head do
+  = render 'devise/registrations/new_sign_up_locale'
+
 %script{src: webpack_asset_path('js/devise/registrations/new_sign_up.js')}
 
-#root
-  = stylesheet_link_tag 'signup', media: 'all'
+#login-type-selection
+  = render 'devise/registrations/new_sign_up_locale'
