+++ dashboard/app/views/layouts/application.html.haml
@@ -46,7 +46,6 @@
     = render inline: File.read(Rails.root.join('..', 'shared', 'haml', 'answerdash.haml')), type: :haml if view_options[:answerdash]
     = javascript_include_tag 'application', 'data-ot-ignore' => ''
     %script{src: webpack_asset_path("js/#{js_locale}/common_locale.js"), 'data-ot-ignore' => ''}
-    %script{src: webpack_asset_path("js/#{js_locale}/signup_locale.js"), 'data-ot-ignore' => ''}
     -# This file is generated by Webpack's SplitChunksPlugin,
     -# which we only use when optimizing our webpack build (ie, non-development environments).
     - if CDO.optimize_webpack_assets
